tictactoe update
checkwinner method almost complete, just have to loop through the 2d array and check if any of the winningPositions exist for a player (to be called in main)

adjusted main to have a massive while loop so code can be reused without having to type prompts or inputs again.

also updated error checking (not really) in the event that a position entered in cin is less than 1 or greater than 9 (not on the board) or enters a position already taken.

a call to place piece (which is modify gameboard provided) is made and the checkwinner call is immediately made.

modifygameboard and checkwinner next on the list