first draft (pending compilation issues)
checkwinner: created a 2d array (or list of lists if you will) of winning positions and used a function signature to declare a generic type of winning position and player position to be used in a containsAll. draws included

determinePlayer: very tacky, just random num from 1 to 100 and if its greater than half its one player's turn to start and its the other if not.

main function: currentPosition and currentPlayer variables used from default initialized var in constructor. updated valid position checker when entering a position. checkwinner implemented immediately after a valid input. swap player feature at the end of the loop.