first debugging gameboard print
solved the gameboard print issue with the player symbols not updating to where they put their position.

positions on the board were marked and updated correctly, as winningpositions were marked and winners or a draw was determined.

reminder: pass by reference vs. pass by  value: value creates a copy so its not printing the real board